Create: Programmers algorithm lessins
There are 𝑛 cargo trucks in a line to cross a one lane wide bridge over a river. The order of trucks cannot be
changed and the weight of each truck may not be the same. Only 𝑤 trucks can be on the bridge at the same
time. We assume that the length of the bridge is 𝑤 unit distance and each truck moves a unit distance in a unit
time. The sum of the weights of the trucks on the bridge should be less than or equal to 𝐿 due to the maximum
safe load of the bridge. Note that the weights of trucks that aren’t fully present on the bridge are not
considered in the load calculation of the trucks on the bridge.